Moved code from PlatformDeviceManager into DeviceManager * Upon further reflection, there didn't really need to be two different device manager classes * Deleted now deprecated PlatformDeviceManager.h/.cpp
