+++ src/psmoveservice/Device/Manager/DeviceManager.cpp
@@ -3,9 +3,11 @@
 
 #include "ControllerManager.h"
 #include "DeviceEnumerator.h"
-#include "PlatformDeviceManager.h"
 #include "HMDManager.h"
 #include "OrientationFilter.h"
+#ifdef WIN32
+#include "PlatformDeviceAPIWin32.h"
+#endif // WIN32
 #include "ServerControllerView.h"
 #include "ServerHMDView.h"
 #include "ServerTrackerView.h"
@@ -17,6 +19,7 @@
 #include "PSMoveProtocol.pb.h"
 #include "PSMoveConfig.h"
 #include "TrackerManager.h"
+
 #include <chrono>
 
 //-- constants -----
@@ -79,15 +82,16 @@ DeviceManager *DeviceManager::m_instance= nullptr;
 
 DeviceManager::DeviceManager()
     : m_config() // NULL config until startup
-#ifdef WIN32
-	, m_platform_manager(new PlatformDeviceManager(_eDevicePlatformApiType_Win32))
-#else
-	, m_platform_manager(new PlatformDeviceManager(_eDevicePlatformApiType_None))
-#endif
+	, m_platform_api_type(_eDevicePlatformApiType_None)
+	, m_platform_api(nullptr)
     , m_controller_manager(new ControllerManager())
     , m_tracker_manager(new TrackerManager())
     , m_hmd_manager(new HMDManager())
 {
+#ifdef WIN32
+	m_platform_api_type = _eDevicePlatformApiType_Win32;
+	m_platform_api = new PlatformDeviceAPIWin32;
+#endif
 }
 
 DeviceManager::~DeviceManager()
@@ -95,7 +99,11 @@ DeviceManager::~DeviceManager()
     delete m_controller_manager;
     delete m_tracker_manager;
     delete m_hmd_manager;
-	delete m_platform_manager;
+
+	if (m_platform_api != nullptr)
+	{
+		delete m_platform_api;
+	}
 }
 
 bool
@@ -111,22 +119,24 @@ DeviceManager::startup()
 	// Save the config back out again in case defaults changed
 	m_config->save();
     
-	success &= m_platform_manager->startup();
-
+	if (m_platform_api != nullptr)
+	{
+		success &= m_platform_api->startup(this);
+	}
 
 	// Register for hotplug events if this platform supports them
 	int controller_reconnect_interval = m_config->controller_reconnect_interval;
 	int tracker_reconnect_interval = m_config->tracker_reconnect_interval;
 	int hmd_reconnect_interval = m_config->hmd_reconnect_interval;
-	if (success && m_platform_manager->get_api_type() != _eDevicePlatformApiType_None)
+	if (success && m_platform_api_type != _eDevicePlatformApiType_None)
 	{
-		m_platform_manager->registerHotplugListener(CommonDeviceState::Controller, m_controller_manager);
+		registerHotplugListener(CommonDeviceState::Controller, m_controller_manager);
 		controller_reconnect_interval = -1;
 
-		m_platform_manager->registerHotplugListener(CommonDeviceState::TrackingCamera, m_tracker_manager);
+		registerHotplugListener(CommonDeviceState::TrackingCamera, m_tracker_manager);
 		tracker_reconnect_interval = -1;
 
-		m_platform_manager->registerHotplugListener(CommonDeviceState::HeadMountedDisplay, m_hmd_manager);
+		registerHotplugListener(CommonDeviceState::HeadMountedDisplay, m_hmd_manager);
 		hmd_reconnect_interval = -1;
 	}
 
@@ -150,7 +160,10 @@ DeviceManager::startup()
 void
 DeviceManager::update()
 {
-	m_platform_manager->poll(); // Send device hotplug events
+	if (m_platform_api != nullptr)
+	{
+		m_platform_api->poll(); // Send device hotplug events
+	}
 
     m_controller_manager->poll(); // Update controller counts and poll button/IMU state
     m_tracker_manager->poll(); // Update tracker count and poll video frames
@@ -173,9 +186,32 @@ DeviceManager::shutdown()
     m_tracker_manager->shutdown();
     m_hmd_manager->shutdown();
 
-	m_platform_manager->shutdown();
+	if (m_platform_api != nullptr)
+	{
+		m_platform_api->shutdown();
+	}
 
     m_instance= nullptr;
+}
+
+// -- Queries ---
+bool 
+DeviceManager::get_device_property(
+	const DeviceClass deviceClass,
+	const int vendor_id,
+	const int product_id,
+	const char *property_name,
+	char *buffer,
+	const int buffer_size)
+{
+	bool bSuccess = false;
+
+	if (m_platform_api != nullptr)
+	{
+		bSuccess = m_platform_api->get_device_property(deviceClass, vendor_id, product_id, property_name, buffer, buffer_size);
+	}
+
+	return bSuccess;
 }
 
 int 
@@ -230,4 +266,51 @@ DeviceManager::getHMDViewPtr(int hmd_id)
     }
 
     return result;
+}
+
+// -- Notification --
+void
+DeviceManager::registerHotplugListener(const CommonDeviceState::eDeviceClass deviceClass, IDeviceHotplugListener *listener)
+{
+	DeviceHotplugListener entry;
+	entry.listener = listener;
+
+	switch (deviceClass)
+	{
+	case CommonDeviceState::Controller:
+	case CommonDeviceState::HeadMountedDisplay:
+		entry.device_class = DeviceClass::DeviceClass_HID;
+		break;
+	case CommonDeviceState::TrackingCamera:
+		entry.device_class = DeviceClass::DeviceClass_Camera;
+		break;
+	default:
+		break;
+	}
+
+	m_listeners.push_back(entry);
+}
+
+void
+DeviceManager::handle_device_connected(enum DeviceClass device_class, const std::string &device_path)
+{
+	for (auto &it = m_listeners.begin(); it != m_listeners.end(); ++it)
+	{
+		if (it->device_class == device_class)
+		{
+			it->listener->handle_device_connected(device_class, device_path);
+		}
+	}
+}
+
+void
+DeviceManager::handle_device_disconnected(enum DeviceClass device_class, const std::string &device_path)
+{
+	for (auto &it = m_listeners.begin(); it != m_listeners.end(); ++it)
+	{
+		if (it->device_class == device_class)
+		{
+			it->listener->handle_device_disconnected(device_class, device_path);
+		}
+	}
 }
