+++ src/psmoveservice/Device/Manager/DeviceManager.h
@@ -2,9 +2,20 @@
 #define DEVICE_MANAGER_H
 
 //-- includes -----
+#include "DeviceInterface.h"
+#include "DevicePlatformInterface.h"
 #include <memory>
 #include <chrono>
-//#include "PSMoveProtocol.pb.h"
+#include <vector>
+
+//-- constants -----
+enum eDevicePlatformApiType
+{
+	_eDevicePlatformApiType_None,
+#ifdef WIN32
+	_eDevicePlatformApiType_Win32,
+#endif // WIN32
+};
 
 //-- typedefs -----
 class DeviceManagerConfig;
@@ -20,13 +31,20 @@ class ServerHMDView;
 typedef std::shared_ptr<ServerHMDView> ServerHMDViewPtr;
 
 //-- definitions -----
+struct DeviceHotplugListener
+{
+	DeviceClass device_class;
+	IDeviceHotplugListener *listener;
+};
+
 /// This is the class that is actually used by the PSMoveService.
-class DeviceManager
+class DeviceManager : public IDeviceHotplugListener
 {
 public:
     DeviceManager();
     virtual ~DeviceManager();
 
+	// -- System ----
     bool startup(); /**< Initialize the interfaces for each specific manager. */
     void update();  /**< Poll all connected devices for each specific manager. */
     void shutdown();/**< Shutdown the interfaces for each specific manager. */
@@ -34,23 +52,45 @@ public:
     static inline DeviceManager *getInstance()
     { return m_instance; }
 
+	// -- Accessors ---
+	ServerControllerViewPtr getControllerViewPtr(int controller_id);
+	ServerTrackerViewPtr getTrackerViewPtr(int tracker_id);
+	ServerHMDViewPtr getHMDViewPtr(int hmd_id);
+
+	// -- Queries ---
+	inline eDevicePlatformApiType get_api_type() const { return m_platform_api_type; }
+	bool get_device_property(
+		const DeviceClass deviceClass,
+		const int vendor_id,
+		const int product_id,
+		const char *property_name,
+		char *buffer,
+		const int buffer_size);
+
     int getControllerViewMaxCount() const;
     int getTrackerViewMaxCount() const;
-    int getHMDViewMaxCount() const;
-        
-    ServerControllerViewPtr getControllerViewPtr(int controller_id);
-    ServerTrackerViewPtr getTrackerViewPtr(int tracker_id);
-    ServerHMDViewPtr getHMDViewPtr(int hmd_id);
+    int getHMDViewMaxCount() const;       
+
+	// -- Notification --
+	void registerHotplugListener(const CommonDeviceState::eDeviceClass deviceClass, IDeviceHotplugListener *listener);
+	void handle_device_connected(enum DeviceClass device_class, const std::string &device_path) override;
+	void handle_device_disconnected(enum DeviceClass device_class, const std::string &device_path) override;
     
 private:
-    DeviceManagerConfigPtr m_config;
+	/// Singleton instance of the class
+	/// Assigned in startup, cleared in teardown
+	static DeviceManager *m_instance;
 
-    /// Singleton instance of the class
-    /// Assigned in startup, cleared in teardown
-    static DeviceManager *m_instance;
+	DeviceManagerConfigPtr m_config;
+
+	// OS specific implementation of hotplug notification
+	eDevicePlatformApiType m_platform_api_type;
+	IPlatformDeviceAPI *m_platform_api;
+
+	// List of registered hot-plug listeners
+	std::vector<DeviceHotplugListener> m_listeners;
 
 public:
-	class PlatformDeviceManager *m_platform_manager;
     class ControllerManager *m_controller_manager;
     class TrackerManager *m_tracker_manager;
     class HMDManager *m_hmd_manager;
